Add a client side with a form, and basic components. Server side additions to create recipe api
Added a react client side, with redux. Added basic components for show; the header component uses redux info to determine if user is logged in and show a log in or log out button. Used redux form for a recipe create form. Added a recipeCreate action creator call the backend and create a recipe to the server.
On the backend added a route to add a recipe to mongodb and a recipe schema. Add a middleware to check if user is logged in.