Change or add lots of little things
Add a source input and display. Move the create button to the top. Display ingredients on a list-item. Check that time on a new recipe form is a number. Display an image on the recipe show page if available. Remove placeholder from header. And some styling tweaks to all components.