Add ability to append a date to a recipe.
Changes to recipelist so that it displays an add button when clicked
from the calendar. The add button invokes an action creator that makes
an ajax call to the server to append a certain date to a recipe. Each
day component is responsible for fetching recipes corresponding to the
day, makes a call on componentDidMount.